requirements.txt
playwright
playwright-stealth
openai
python-dotenv
pydantic
beautifulsoup4
tenacity
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
